# Adding OpenACC statements to accelerate MYNN surface scheme performance through GPU offloading