@dustinswales 2023-08-21
@timsliwinski-noaa Thanks for this contribution. 
A couple comments/questions.
- I'm not sure the code works as is "inline using GPUs". See my comment on the changes you made to errmsg in the code. We don't test the GPU functionality as part of our testing system, which is something we are working on, but I'm pretty sure the pre-processor directives around a local declaration of a host-side interstitial might break things. As a workaround you could declare a local copy of errmsg, with explicit extent, that you pass to the GPU.
- Do we need this new pre-processor directive, __OPENACC, which is acting to bypass the physics to host-model communication for displaying debugging information? Maybe this could be resolved with some changes in the scheme.
- I don't see pull-requests into the ccpp-physics supermodules [fv3atm](https://github.com/NOAA-EMC/fv3atm/pulls) or the [UFS weather model](https://github.com/ufs-community/ufs-weather-model/pulls). Have these changes been tested in the UFS?