@timsliwinski-noaa 2023-08-21
Hi @dustinswales and @joeolson42. Thanks for your feedback. I'm reviewing your suggestions and looking at alternatives. I did want to address some things I didn't mention in my initial PR that may be helpful based on your responses:

First off, I want to mention I approached this effort with an explicit goal of minimizing overall code changes. This did limit what I felt I could change, but was done for the sake of code maintainers being able to recognize their original scientific code and leaving the scheme more or less as intact as when I started. If there is willingness to engage in limited invasive code changes, I'm open to that discussion if you are as it probably would reduce the amount of preprocessor directives scattered about.

Also, since I didn't mention testing -- other than performance testing -- in much detail in my initial PR (apologies!), I wanted to add that right away:

Testing was done, but not with the UFS. Instead, the CCPP Single Column Model was used for testing, though indirectly. Input to and output from the MYNN surface scheme from single column model runs was captured and fed back into a standalone version of the scheme isolated with a custom driver. That driver was built just to read data in and output the results for more lightweight testing and accuracy of solution evaluation. Overall, the expected output was achieved with nothing more than slight differences in last-digit rounding for real numbers as would be expected based on differences between GPU and CPU floating point computations.

Having said that, if there is a proper test case to run the full UFS or CCPP framework with, I would gladly test with those to see how these changes work as a drop-in replacement. As of right now, I was unaware testing of CCPP physics with the full UFS was necessary. I was under the impression that CCPP was meant to be model independent (is that wrong?). I'm also unaware of any established baselines to test physics accuracy of solution against. I am open to suggestions if there are testing best practices I'm missing and should incorporate.

Thanks, and I will return with more information for your specific review points soon after I look at them more closely again.