@timsliwinski-noaa 2023-08-24
I'm pushing up a new change to this branch that hopefully addresses the variable substitution by preprocessor directive issue. However, some changes still require preprocessor directives to be used based on the __OPENACC compiler variable. Please see the commit message for more details.

Next, I'm going to look more closely at trying to re-enable some of the debug output print statements, but didn't want to intermingle those changes with this latest commit unless the addition of new, extra errmsg variables sparked more concerns.