@timsliwinski-noaa 2023-08-28
@joeolson42 and @dustinswales -- I have pushed up the final round of requested changes, this time regarding the disabled debugging output. The output has been re-enabled, and a few data optimizations and use of the OpenACC serial construct were necessary to accomplish this. Overall, though, performance was not majorly impacted (except when debugging is enabled) because the majority of the IO operations were already optimized. There is a pretty big hit to compilation time with debugging turned on since the OpenACC compiler must try to optimize a lot of things that are not well-suited for the accelerator, so keep that in mind when trying to test with debugging enabled. Further details are in the commit message. Please let me know if you have any other questions, comments, or suggested changes.