@grantfirl 2023-08-30
It seems like we should use this PR as an opportunity to define some best practices with respect to CCPP error handling and debugging print statements when running on GPUs. I don't think that it would be crazy, for example, to declare errmsg with an explicit length. I don't think that doing this with preprocessors directives (like using __OPENACC) would work because of the CCPP prebuild check on the "kind" attribute, but just redefining errmsg as len=256 or something permanently (whether or not we're using GPUs) seems reasonable to me.

@timsliwinski-noaa What is the preferred way to handle debugging write statements with GPU-accelerated code? Do we just need to make sure that we're not branching inside an accelerated section? E.g. if we wanted to conditionally write out anything within an accelerated loop, do we need to just redo the same loop on the CPU side when a debug flag is active or something?

Also, I have a question about the "optimized data movement". You're referring to all `acc enter/exit data copyX` calls being done in the calling layer of the top-level scheme, right? For example, for mynnsfc_wrapper_run, we would want all of the data copying openacc calls in the CCPP autogenerated cap, right?